+++ test-e2e/tests/dns.test.ts
@@ -1,30 +1,28 @@
 import got from 'got';
 import { expect } from 'chai';
-import { waitMesurementFinish } from './utils.js';
+import { waitMesurementFinish } from '../utils.js';
 
-describe('http mesurement', () => {
+describe('dns mesurement', () => {
 	it('should finish successfully', async () => {
 		const { id } = await got.post('http://localhost:3000/v1/measurements', { json: {
 			target: 'www.jsdelivr.com',
-			type: 'http',
+			type: 'dns',
 		} }).json();
 
 		const { response, body } = await waitMesurementFinish(id);
 
 		expect(body.status).to.equal('finished');
 		expect(body.results[0].result.status).to.equal('finished');
-		expect(body.results[0].result.rawBody).to.equal(null);
+		expect(body.results[0].result.hops).to.not.exist;
 		expect(response).to.matchApiSchema();
 	});
 
-	it('should finish successfully in case of GET request', async () => {
+	it('should finish successfully in case of "traced": true', async () => {
 		const { id } = await got.post('http://localhost:3000/v1/measurements', { json: {
 			target: 'www.jsdelivr.com',
-			type: 'http',
+			type: 'dns',
 			measurementOptions: {
-				request: {
-					method: 'GET',
-				},
+				trace: true,
 			},
 		} }).json();
 
@@ -32,7 +30,7 @@ describe('http mesurement', () => {
 
 		expect(body.status).to.equal('finished');
 		expect(body.results[0].result.status).to.equal('finished');
-		expect(body.results[0].result.rawBody.length > 0).to.be.true;
+		expect(body.results[0].result.hops.length > 0).to.be.true;
 		expect(response).to.matchApiSchema();
 	});
 });
